+++ src/rebar3_grisp_build.erl
@@ -146,7 +146,7 @@ event([build, validate, apps, {grisp_dir_without_dep, A}]) ->
 event([build, validate, version]) ->
     io:format("* Resolving OTP version~n");
 event([build, validate, version, {selected, Version, Target}]) ->
-    io:format("    Selected version ~s (requirement was ~s)~n", [Version, Target]);
+    io:format("    ~s (requirement was \"~s\")~n", [Version, Target]);
 event([build, collect, {hash, Hash, Index}]) ->
     debug("GRiSP hash:~n~s~n~n~p", [Hash, Index]);
 event([build, download]) ->