+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2022-02-16
+
 ### Added
 
 - `rebar3 grisp package list` can now take a `--hash` flag that prints all hash versions of a package
@@ -237,7 +239,8 @@ Modify release configuration dynamically
 
 * Initial release.
 
-[unreleased]: https://github.com/grisp/rebar3_grisp/compare/2.1.0...HEAD
+[unreleased]: https://github.com/grisp/rebar3_grisp/compare/2.2.0...HEAD
+[2.2.0]: https://github.com/grisp/rebar3_grisp/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/grisp/rebar3_grisp/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/grisp/rebar3_grisp/compare/1.3.0...2.0.0
 [1.3.0]: https://github.com/grisp/rebar3_grisp/compare/1.2.6...1.3.0